% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Sybil}
\alias{predict.Sybil}
\title{Predict Series using Sybil}
\arguments{
\item{object}{An \code{S7} object with class \code{"Sybil"}.}

\item{series}{An \code{S7} object with class \code{"Series"}.}

\item{return_attentions}{Logical. Whether to return attentions?
Default is \code{FALSE}.}

\item{threads}{Number of CPU threads to use for PyTorch inference. Default
is \code{0L}.}

\item{...}{Not currently used.}
}
\value{
An \code{S7} object with properties \verb{@scores}, and \verb{@attentions}.
\verb{@attentions} will be an empty list unless \code{return_attentions} was \code{TRUE}.
}
\description{
For a series or list of series, estimate lung cancer risk at 1 to 6 years
and optionally attentions maps.
}
