% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Series.R
\name{Series}
\alias{Series}
\alias{has_label}
\title{Create a Series}
\usage{
Series(
  dicoms = character(0),
  voxel_spacing = integer(0),
  label = integer(0),
  censor_time = integer(0),
  file_type = "dicom",
  split = "test"
)

has_label(x)
}
\arguments{
\item{dicoms}{Character vector of DICOM files.}

\item{voxel_spacing}{Optional. Numeric vector of voxel spacings associated
with the input CT as (row spacing, col spacing, slice thickness).}

\item{label}{Optional. Label for whether the individual has or ever developed
lung cancer (0=No, 1=Yes).}

\item{censor_time}{Optional. Number of years until the lung cancer diagnosis.
If less than 1 year, set to \code{0}. Must be present if \code{label} is provided.}

\item{file_type}{Optional. File type of CT slices. One of \code{"png"} or
\code{"dicom"}. Default is \code{"dicom"}.}

\item{split}{Optional. Data split that the series belongs to. One of
\code{"train"}, \code{"dev"}, or \code{"test"}. Default is \code{"test"}.}

\item{x}{An \code{S7} object of class \code{"Series"}.}
}
\value{
An \code{S7} object of class \code{"Series"}.
}
\description{
Initialize a Series to estimate lung cancer risk.
}
