% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Sybil.R
\name{Sybil}
\alias{Sybil}
\alias{load_model}
\alias{Sybil_base}
\alias{Sybil_1}
\alias{Sybil_2}
\alias{Sybil_3}
\alias{Sybil_4}
\alias{Sybil_5}
\alias{Sybil_ensemble}
\title{Create a Sybil model object.}
\usage{
Sybil(
  name_or_path = "sybil_ensemble",
  cache = "~/.sybil/",
  calibrator_path = character(0),
  device = character(0)
)

load_model(name_or_path, cache, calibrator_path, device)

Sybil_base(cache = NULL, device = NULL)

Sybil_1(cache = NULL, device = NULL)

Sybil_2(cache = NULL, device = NULL)

Sybil_3(cache = NULL, device = NULL)

Sybil_4(cache = NULL, device = NULL)

Sybil_5(cache = NULL, device = NULL)

Sybil_ensemble(cache = NULL, device = NULL)
}
\arguments{
\item{name_or_path}{Alias to a provided pretrained Sybil model or path
to a Sybil checkpoint. Names must be one of \code{"sybil_base"}, \code{"sybil_1"},
\code{"sybil_2"}, \code{"sybil_3"}, \code{"sybil_4"}, \code{"sybil_5"}, or \code{"sybil_ensemble"}.
Default is \code{"sybil_ensemble"}.}

\item{cache}{Directory to download model checkpoints. Default is in \verb{.sybil/}
in the users home directory (\code{~}).}

\item{calibrator_path}{Optional. Path to calibrator pickle file corresponding
with the chosen model.}

\item{device}{Optional. If provided, will run inference using this device. By
default, uses GPU with the most free memory, if available.}
}
\value{
An \code{S7} object of class \code{"Sybil"}.
}
\description{
Load a Sybil model to estimate lung cancer risk at 1 to 6 years. The Sybil
model weights will be downloaded to disk on first use (to \code{cache}). More
information about the Sybil model can be found here:
\url{https://doi.org/10.1200/jco.22.01345}.
}
\examples{
model <- Sybil()
}
