% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sybil-package.R
\docType{package}
\name{sybilr-package}
\alias{sybilr}
\alias{sybilr-package}
\title{sybilr: Lung Cancer Risk Preduction with Sybil}
\description{
Provides an R interface to 'Sybil', a Python-based neural network for lung cancer prediction from low-dose computed tomography (LDCT) scans. This package enables users to run Sybil models and integrate LDCT-based lung cancer risk prediction workflows within R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattwarkentin/sybilr}
  \item Report bugs at \url{https://github.com/mattwarkentin/sybilr/issues}
}

}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{matthew.warkentin@ucalgary.ca} [copyright holder]

}
\keyword{internal}
