% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate Series using Sybil}
\usage{
evaluate(object, series, return_attentions = FALSE, ...)
}
\arguments{
\item{object}{An \code{S7} object with class \code{"Sybil"}.}

\item{series}{An \code{S7} object with class \code{"Series"}.}

\item{return_attentions}{Logical. Whether to return attentions?
Default is \code{FALSE}.}

\item{...}{Not currently used.}
}
\value{
An \code{S7} object with properties \verb{@auc}, \verb{@c_index}, \verb{@scores}, and
\verb{@attentions}, if \code{return_attentions} was \code{TRUE}.
}
\description{
For a list of labelled series, estimate lung cancer risk predictions (and
optionally attentions) and the C-index and time-dependent area
under the receiver operating characteristic curve (AUC) model performance
metrics.
}
