% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Sybil}
\alias{predict.Sybil}
\title{Predict a Series using Sybil}
\arguments{
\item{object}{An object with class \code{"Sybil"}.}

\item{series}{An object with class \code{"Series"}.}

\item{return_attentions}{Logical. Whether to return attentions?
Default is \code{FALSE}.}

\item{threads}{Number of CPU threads to use for PyTorch inference. Default
is \code{0L}.}
}
\value{
An \code{S7} object with properties \verb{@scores}, and \verb{@attentions}, if
\code{return_attentions} was \code{TRUE}.
}
\description{
For a list of \code{Series}, predict...
}
