% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize-attentions.R
\name{visualize_attentions}
\alias{visualize_attentions}
\title{Visualizae Attentions Maps}
\usage{
visualize_attentions(series, attentions, dir = NULL, gain = 3)
}
\arguments{
\item{series}{List of \code{Series}.}

\item{attentions}{List of attentions.}

\item{dir}{Directory to save GIFs.}

\item{gain}{Gain. Default is 3.}
}
\value{
Images with attentions overlaid on the original series if \code{dir} is
not set, otherwise returns \code{dir}, invisibly.
}
\description{
Visualizae Attentions Maps
}
